template_service/pyproject.toml
[project]
name = "distributed-notification-system"
version = "0.1.0"
description = "Template service for distributed-notification-system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-ninja>=1.4.5",
    "django-redis>=6.0.0",
    "django-stubs>=5.2.7",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.12",
    "pytest>=9.0.0",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "uvicorn>=0.38.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
fix = true
select = ["E", "F", "I"] 
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
